Make sure we wait for the swap has finished before we launch the game
Since all we do is rename folders, 3 minutes should be more than enough time even on hard drives, but just to be sure, we WON'T launch the game if it fails to do so in the time allotted. We also warn the player they might want to reinstall the game just to be safe.